pyproject.toml
[tool.poetry]
name = "poetry"
version = "0.1.0"
description = ""
authors = ["Antti Nevalainen <antti.s.nevalainen@helsinki.fi>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.20.2"
pandas = "^1.2.3"
invoke = "^1.5.0"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
